requirements.txt
Django==4.0.6
django-extensions
ipython
django_admin_rangefilter
celery~=5.2.7
django-filter
djangorestframework
drf-excel
djangorestframework-simplejwt
drf-yasg


# dev
django-debug-toolbar==3.5.0
django-silk

# Flake8
flake8==4.0.1
flake8-builtins==1.5.3
flake8-print==5.0.0

#tests
pytest
pytest-django
pytest-cov
pytest-mock